feat: use graphql query instead of rest calls
you might ask why the expected result of the happy path test changed:
I had simplified the output/response of the events call to the rick and morty api to avoid tons of calls that i'd have needed to mock (for every character). apparently, this shows the downsides of tampering with test data.